Avoid direct imports from __init__
